chore: basic js-ts file rename
