+++ src/js/core/audio-functions.ts
