+++ src/js/markup/slider.ts
