--- src/js/outputs/index.js
@@ -1,3 +1,0 @@
-import { picobel } from "../Picobel.js";
-
-export default picobel;
