+++ src/js/outputs/index.ts
@@ -1,0 +1,3 @@
+import { picobel } from "../Picobel";
+
+export default picobel;
