--- src/js/outputs/legacy.js
@@ -1,3 +1,0 @@
-import { picobel as Picobel } from "../Picobel.js";
-
-window.picobel = Picobel;
