+++ src/js/outputs/legacy.ts
@@ -1,0 +1,3 @@
+import { picobel as Picobel } from "../Picobel";
+
+window.picobel = Picobel;
