+++ src/js/wc/web-component.default.ts
@@ -1,4 +1,4 @@
-import { picobel } from "../Picobel.js";
+import { picobel } from "../Picobel";
 
 import styles from "../../../build/picobel.default.css";
